+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidOptionsInitializer.java
@@ -389,6 +389,9 @@ final class AndroidOptionsInitializer {
       options.addIntegration(replay);
       options.setReplayController(replay);
     }
+    options
+        .getFeedbackOptions()
+        .setDialogHandler(new SentryAndroidOptions.AndroidUserFeedbackIDialogHandler());
   }
 
   /**