+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -3035,7 +3035,11 @@ public class SentryOptions {
     final @NotNull SdkVersion sdkVersion = createSdkVersion();
     experimental = new ExperimentalOptions(empty, sdkVersion);
     sessionReplay = new SentryReplayOptions(empty, sdkVersion);
-    feedbackOptions = new SentryFeedbackOptions();
+    feedbackOptions =
+        new SentryFeedbackOptions(
+            configurator ->
+                logger.log(SentryLevel.WARNING, "showDialog() can only be called in Android."));
+
     if (!empty) {
       setSpanFactory(SpanFactoryFactory.create(new LoadClass(), NoOpLogger.getInstance()));
       // SentryExecutorService should be initialized before any